Car Mount
A design I made to mount a custom computer in the dash of my Corolla. It will function as an adapter between the old mount and the new mount, to use the clip mechanisms for the stock computer but provide a surface in which to drill holes to screw the new computer into.

The 3mf file is included because a wide array of modifications to the print settings was required to get it to work. This may need updating in the near future, the design is still printing now and may not work.